requirements.txt
aiofiles
aiohttp
apscheduler
aria2p
cairosvg
cloudscraper
colour
cowpy
dataclasses
DateTime
emoji
emoji-country-flag
fake_useragent
fonttools
geopy
gitpython
glitch_this
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
googletrans==4.0.0-rc1
gtts
hachoir
herokum
html-telegraph-poster
humanize
IMDbPY
imutils
instaloader
jikanpy
justwatch
lottie
lyricsgenius
markdown
motor
moviepy
nekos.py
opencv-python-headless
Pillow
prettytable
psutil
psycopg2
PyDictionary
pyfiglet
PyGithub
pygments
pylast
pymediainfo
PyMuPDF
pySmartDL
python-barcode
python-dotenv
pytz
qrcode
randomstuff.py
regex
requests
scipy
search-engine-parser
selenium
ShazamAPI
spamwatch
speedtest-cli
spotify_token
sqlalchemy-json
sqlalchemy==1.3.24
telegraph
git+https://github.com/MUTLCC/Telethon
tgcrypto
ujson
urlextract
validators
vcsi
wget
git+https://github.com/TeleDoge/Wikipedia
wordcloud
youtube-search-python
youtube_dl